Extracted Google connection test
