+++ tests/CMakeLists.txt
@@ -5,3 +5,4 @@ enable_testing()
 include_directories(${CMAKE_CURRENT_SOURCE_DIR})
 
 add_subdirectory(ChunkData)
+add_subdirectory(Network)
